deleteCategory action added to CategoryController, also deleteCategoryById method implemented in CategoryService

+++ src/main/java/pl/coderslab/charity/category/CategoryService.java
@@ -24,6 +24,10 @@ public class CategoryService {
     }
 
     public void deleteCategoryById(Long id) {
+        boolean exists = categoryRepository.existsById(id);
+        if (!exists) {
+            throw new IllegalStateException("Category with given id does not exist");
+        }
         categoryRepository.deleteById(id);
     }
 }
